Add option to return the timecode that matched when verifying TOTP.
This adds a boolean option `return_timecode` to TOTP.verify that
changes its return value on success to an integer containing the
timecode that was used to generate the matching OTP.

This is especially useful when `valid_window` is not 0, as then
several values might be valid at the same moment. If an implementation
compares just the bare OTP, replaying a recent value is possible.
By comparing timecodes, that can be prevented, as they are strictly
increasing.

The patch also fixes a minor typing issue, as the `for_time` argument
can also be an number - there even is a test that calls it like this.

Finally, the special case for `valid_window == 0` has been removed as
it is a micro-optimization that only leads to repeated code.