typing: accept float for timestamp values
One gotcha to be aware of: Python type checking is special with
floats (https://typing.python.org/en/latest/spec/special-types.html#special-cases-for-float-and-complex)
so annotating a variable as float allows passing in an int. However,
when doing runtime checks, code needs to test for
`isinstance(..., (float, int))`

Tests were already passing -29.5 to `TOTP.at`; note that casting a
float to an int in Python just strips off the fractional part (like
`floor(3)`). Because TOTP intervals are integers, this means that this
rounding off operation does not change which interval the timestamp is
in so we can do this safely.