typing: accept float for timestamp values
Tests were already passing -29.5 to `TOTP.at`; note that casting a
float to an int in Python just strips off the fractional part (like
`floor(3)`). Because TOTP intervals are integers, this means that this
rounding off operation does not change which interval the timestamp is
in so we can do this safely.